+++ setup.sh
@@ -60,10 +60,18 @@ A Service Principal will be created in order to give access to the Azure resourc
 This operation needs Azure Active Directory privilege for creating AAD application.
 After creating the Service Principal, you will be asked on which subscription the access should be given.
 
-${bldora}/!\ ⚠ If you input the name of an existing service principal, the existing one will be used instead of creating a new one but the secret will be reset. ⚠ /!\\
+${bldora}/!\ ⚠ If you input the name of an existing service principal, the existing one will be used instead of creating a new one. ⚠ /!\\
 ${txtrst}
 
 EOT
+
+function create_az_sp() {
+  # TODO manage fails
+  SP_RESULT=$(az ad sp create-for-rbac -n "$SP_NAME" --skip-assignment --query "join('#', [appId,password])" -o tsv)
+  SP_APP_ID=$(echo "$SP_RESULT" | cut -f1 -d'#')
+  SP_APP_SECRET=$(echo "$SP_RESULT" | cut -f2 -d'#')
+  SP_OBJECT_ID=$(az ad sp show --id "$SP_APP_ID" --query 'objectId' -o tsv)
+}
 
 # Create Service Principal
 INPUT_SPNAME="_"
@@ -75,13 +83,22 @@ done
 SP_NAME=${INPUT_SPNAME:-$DEFAULT_SPNAME}
 
 printf "\n"
-echo "Creating Service Principal \"$SP_NAME\""
-# TODO manage fails
-SP_RESULT=$(az ad sp create-for-rbac -n "$SP_NAME" --skip-assignment --query "join('#', [appId,password])" -o tsv)
-SP_APP_ID=$(echo "$SP_RESULT" | cut -f1 -d'#')
-SP_APP_SECRET=$(echo "$SP_RESULT" | cut -f2 -d'#')
-SP_OBJECT_ID=$(az ad sp show --id "$SP_APP_ID" --query 'objectId' -o tsv)
-echo "Done creating Service Principal with id $SP_APP_ID"
+echo "Checking if Service Principal \"$SP_NAME\" already exists"
+SP_APP_ID=$(az ad sp list --query "[?displayName=='$SP_NAME'].appId" -o tsv)
+if [ -z "$SP_APP_ID" ]; then
+  echo "Service Principal \"$SP_NAME\" not found"
+  echo "Creating Service Principal \"$SP_NAME\""
+  create_az_sp
+  echo "Done creating Service Principal with id $SP_APP_ID"
+else
+  echo "Service Principal \"$SP_NAME\" found with AppId \"$SP_APP_ID\""
+  read -n 1 -r -p "Do you want to reset the password of the current Service Principal \"$SP_NAME\"($SP_APP_ID) (y/N): " RESETPWD
+  if [[ "${RESETPWD,,}" = 'y' ]]; then
+    echo "Resetting Service Principal \"$SP_NAME\" password"
+    create_az_sp
+    echo "Done resetting Service Principal with id $SP_APP_ID"
+  fi
+fi
 
 cat <<EOT